+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove unused code related to auto-open record detail view. Refs UIIN-2819.
 * Replace `all` with the `=` operator to get correct results when using the `All` search option. Refs UIIN-2816.
 * Keyboard shortcuts modal: Add quickMARC shortcuts to modal. Refs UIIN-2795.
+* Create new settings for classification type sorting. Refs UIIN-2775.
 * Add "Display to public" column to receiving history on Holdings. Refs UIIN-2826.
 
 ## [11.0.1] (IN PROGRESS)